requirements.txt
logzero
dynaconf
pymitter
paho-mqtt
